Brought code up to demo level, using static API
