+++ README.md
@@ -1,22 +1,34 @@
-# Lab: Class 39: React III
+# Lab: Class 39: React III - Next.js
 
 ## Open Git Pull Requests  
 
 
 ## Overview  
 
+Build a multi page web app using Next.js to consume an external API.
+
+The app will display most recent XKCD comic and have links to previous 10 comics.
 
 ## Feature Tasks and Requirements  
-
+- Refer to [example site](https://xkcd-two.now.sh/)  
+- Web app should have 2 static routes
+    - `Home` page at `“/”`
+    - `About` page at `“/about”`
+- Web app should have dynamic routes
+    - Dynamic route should display a single comic
 
 ## Implementation Notes:  
-
+- Refer to [XDCD API](https://xkcd.com/json.html)  
+    - Inspect the data returned to determine the dynamic links needed.
+- Remember to edit package.json script section
 
 ## User Acceptance Tests  
-
+- No automated testing today. Your goal is to match appearance and functionality of reference web app.
 
 ## Dependencies  
-
+- "next": "9.4.4",
+- "react": "16.13.1",
+- "react-dom": "16.13.1"
 
 ## Authors  
 - Software Developer: Joseph Zabaleta
@@ -29,7 +41,8 @@
 This project is under the MIT License.
 
 ## Acknowledgements / Resources  
-- none
+- [Example Website](https://xkcd-two.now.sh/)  
+- [XKCD API](https://xkcd.com/json.html)  
 
 ## Version History  
 - 1.0.0 20200702
