+++ components/Nav.js
@@ -1,0 +1,18 @@
+import Link from 'next/link'
+
+export default function Nav() {
+    return (
+        <ul>
+            <li>
+                <Link href="/">
+                    <a>Home</a>
+                </Link>
+            </li>
+            <li>
+                <Link href="/about">
+                    <a>About</a>
+                </Link>
+            </li>
+        </ul>
+    )
+}
