+++ pages/about.js
@@ -1,0 +1,10 @@
+import Nav from '../components/Nav'
+
+export default function About() {
+    return (
+        <>
+            <Nav />
+            <h1>About Page</h1>
+        </>
+    )
+}
